+++ sdk/identity/identity/src/credentials/onBehalfOfCredential.browser.ts
@@ -1,0 +1,21 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { TokenCredential, AccessToken } from "@azure/core-auth";
+
+import { credentialLogger, formatError } from "../util/logging";
+
+const BrowserNotSupportedError = new Error("OnBehalfOfCredential is not supported in the browser.");
+const logger = credentialLogger("OnBehalfOfCredential");
+
+export class OnBehalfOfCredential implements TokenCredential {
+  constructor() {
+    logger.info(formatError("", BrowserNotSupportedError));
+    throw BrowserNotSupportedError;
+  }
+
+  public getToken(): Promise<AccessToken | null> {
+    logger.getToken.info(formatError("", BrowserNotSupportedError));
+    throw BrowserNotSupportedError;
+  }
+}